+++ samples/microui/build.bat
@@ -15,7 +15,7 @@ set wasmFlags=--target=wasm32^
        -I%ORCA_DIR%/src/ext
 
 :: build sample as wasm module and link it with the orca module
-clang %wasmFlags% -L %ORCA_DIR%/bin -lorca_wasm -o module.wasm src/main.c microui/microui.c
+clang %wasmFlags% -L %ORCA_DIR%/bin -lorca_wasm -o module.wasm src/main.c src/microui/microui.c
 IF %ERRORLEVEL% NEQ 0 EXIT /B %ERRORLEVEL%
 
 :: create app directory and copy files into it
